move deploy_service to pipeline utils
