+++ launch/pipeline/__init__.py
@@ -1,11 +1,12 @@
 from launch.pipeline.deployment import Deployment
 from launch.pipeline.runtime import Runtime
 from launch.pipeline.service import (
-    ServiceDescription,
     SequentialPipelineDescription,
+    ServiceDescription,
     SingleServiceDescription,
 )
 from launch.pipeline.utils import (
+    deploy_service,
     find_all_sub_service_descriptions,
     make_sequential_pipeline,
     make_service,
